% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{AlignPeakParam}
\alias{AlignPeakParam}
\title{Parameters for aligning MS1 peaks}
\usage{
AlignPeakParam(
  mz_tol = 10,
  strict_rt_constrains = TRUE,
  rt_tol_landmark = 30,
  rt_tol_match = 15,
  ccs_tol = 1.5,
  cutoff = 0.8,
  min_num_fragments = 4,
  method_match = c("direct", "bootstrap"),
  method_score = c("dp", "msdial", "combined"),
  weight_intensity = 1,
  weight_mz = 0,
  int_threshold_abs = 0,
  int_threshold_rel = 0,
  include_precursor = TRUE,
  mz_tol_ms1 = 20,
  mz_tol_ms2 = 35,
  rerun = FALSE
)
}
\arguments{
\item{mz_tol}{\code{numeric} m/z tolerance for binning precursor ions}

\item{strict_rt_constrains}{\code{logical} if remove peaks out of defined rt range in the experiment}

\item{rt_tol_landmark}{\code{numeric} retention time tolerance for finding landmarks}

\item{rt_tol_match}{\code{numeric} retention time tolerance for finding matched peaks}

\item{ccs_tol}{\code{numeric} CCS tolerance for binning precursor ions in percentage}

\item{cutoff}{\code{numeric} cutoff value for defing a precursor ion bin}

\item{method_match}{\code{character} method for matching MSMS spectra}

\item{method_score}{\code{character} method for scoring MSMS spectra}

\item{weight_intensity}{\code{numeric} weight of intensity when scoring}

\item{weight_mz}{\code{numeric} weight of m/z when scoring}

\item{int_threshold_abs}{\code{numeric} minimal absolute intensity to be kept in MSMS spectra}

\item{int_threshold_rel}{\code{numeric} minimal relative intensity to be kept in MSMS spectra}

\item{include_precursor}{\code{logical} if include precursor ion in MSMS spectra when matching}

\item{mz_tol_ms1}{\code{numeric} m/z tolerance for defining a precursor ion}

\item{mz_tol_ms2}{\code{numeric} m/z tolerance to define a matched MSMS spectra in msms matching}

\item{rerun}{\code{logical} if rerun the process}

\item{weight_rt}{\code{numeric} weight for retention time to calculate precursor ion distance}

\item{weight_mobility}{\code{numeric} weight for ion mobility to calculate precursor ion distance}

\item{weight_msms}{\code{numeric} weight for MSMS intensity to calculate precursor ion distance}
}
\description{
Parameters for aligning MS1 peaks
}
