% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{BinPrecursorParam}
\alias{BinPrecursorParam}
\title{Parameters for binning precursor ions}
\usage{
BinPrecursorParam(
  mz_tol = 20,
  rt_tol = c(10, 20),
  mobility_tol = c(0.015, 0.03),
  distance_cutoff = 1,
  weight_rt = 1,
  weight_mobility = 1,
  weight_msms = 1,
  method_match = c("direct", "bootstrap"),
  method_score = c("dp", "msdial", "combined"),
  weight_intensity = 1,
  weight_mz = 0,
  int_threshold_abs = 0,
  int_threshold_rel = 0,
  include_precursor = TRUE,
  mz_tol_ms1 = 20,
  mz_tol_ms2 = 35,
  rerun = FALSE
)
}
\arguments{
\item{mz_tol}{\code{numeric} m/z tolerance for binning precursor ions}

\item{rt_tol}{\code{numeric} retention time tolerance for binning precursor ions}

\item{mobility_tol}{\code{numeric} ion mobility tolerance for binning precursor ions}

\item{distance_cutoff}{\code{numeric} cutoff value for defing a precursor ion bin}

\item{weight_rt}{\code{numeric} weight for retention time to calculate precursor ion distance}

\item{weight_mobility}{\code{numeric} weight for ion mobility to calculate precursor ion distance}

\item{weight_msms}{\code{numeric} weight for MSMS intensity to calculate precursor ion distance}

\item{method_match}{\code{character} method for matching MSMS spectra}

\item{method_score}{\code{character} method for scoring MSMS spectra}

\item{weight_intensity}{\code{numeric} weight of intensity when scoring}

\item{weight_mz}{\code{numeric} weight of m/z when scoring}

\item{int_threshold_abs}{\code{numeric} minimal absolute intensity to be kept in MSMS spectra}

\item{int_threshold_rel}{\code{numeric} minimal relative intensity to be kept in MSMS spectra}

\item{include_precursor}{\code{logical} if include precursor ion in MSMS spectra when matching}

\item{mz_tol_ms1}{\code{numeric} m/z tolerance for defining a precursor ion}

\item{mz_tol_ms2}{\code{numeric} m/z tolerance to define a matched MSMS spectra in msms matching}

\item{rerun}{\code{logical} if rerun the process}
}
\description{
Parameters for binning precursor ions
}
