% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{CombineParam}
\alias{CombineParam}
\title{Combine score parameter setup}
\usage{
CombineParam(
  cutoff = 0.6,
  weightRT = 0.2,
  weightCCS = 0.4,
  weightMSMS = 0.4,
  scoreMSMS = c("reverse", "forward")
)
}
\arguments{
\item{cutoff}{\code{numeric} threshold for a avaliable match of combined score}

\item{weightRT}{\code{numeric} weight of RT when combining the scors}

\item{weightCCS}{\code{numeric} weight of CCS when combining the scors}

\item{weightMSMS}{\code{numeric} weight of MSMS when combining the scors}

\item{scoreMSMS}{\code{character} MSMS score used for score combination
(only "reverse" or "forward" is supported)}
}
\value{
a \code{CombineParam} object
}
\description{
Combine score parameter setup
}
