% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{DereplicatePeaksParam}
\alias{DereplicatePeaksParam}
\title{Parameters for dereplicate MS1 peaks}
\usage{
DereplicatePeaksParam(
  order_column = "area",
  mz_tol = 20,
  rt_tol = 20,
  mobility_tol = 0.015,
  match_msms = FALSE,
  msms_cutoff = 0.8,
  method_match = c("direct", "bootstrap"),
  method_score = c("dp", "msdial", "combined"),
  weight_intensity = 1,
  weight_mz = 0,
  int_threshold_abs = 0,
  int_threshold_rel = 0,
  include_precursor = TRUE,
  mz_tol_ms1 = 20,
  mz_tol_ms2 = 35,
  rerun = FALSE
)
}
\arguments{
\item{mz_tol}{\code{numeric} m/z tolerance for binning precursor ions}

\item{rt_tol}{\code{numeric} retention time tolerance for binning precursor ions}

\item{mobility_tol}{\code{numeric} ion mobility tolerance for binning precursor ions}

\item{rerun}{\code{logical} if rerun the process}
}
\description{
Parameters for dereplicate MS1 peaks
}
