% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_6560.R
\name{.update_ioi_with_calibration_table}
\alias{.update_ioi_with_calibration_table}
\title{update ioi table with calibration table}
\usage{
.update_ioi_with_calibration_table(file_name, calibration_table, ioi_table)
}
\description{
update ioi table with calibration table
}
