% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Experiment.R
\name{Experiment}
\alias{Experiment}
\title{Params for MS experiment}
\usage{
Experiment(
  wd = ".",
  experiment_type = c("DDA", "DIA"),
  injection_order = NULL,
  ion_mode = c("positive", "negative"),
  res_dir = "results",
  tmp_dir = file.path(res_dir, "tmp"),
  ce = "30",
  ms1range = NULL,
  ms2range = NULL,
  rt_range = c(0, 720),
  lc_column = c("RP", "HILIC"),
  lc_method = "RP18",
  res_define_at = 200,
  nSlaves = 4,
  BPPARAM = NULL
)
}
\arguments{
\item{wd}{\code{character} Working directory for experiment to be processed}

\item{injection_order}{\code{character} Injection order file name (relative to wd).
Two/Three columns are required: 'file', 'order', (optional) 'sample_type'}

\item{ion_mode}{\code{charactor} ionization mode for MS experiment data acquisition}

\item{res_dir}{\code{character} directory for saving results}

\item{tmp_dir}{\code{character} directory for saving temporary files}

\item{ce}{\code{character} CE for MS experiment}

\item{ms1range}{\code{numeric(2) or NULL} Mass range setup for MS1 acquisition}

\item{ms2range}{\code{numeric(2) or NULL} Mass range setup for MS2 acquisition}

\item{lc_column}{\code{character} LC column for MS experiment}

\item{lc_method}{\code{character} LC method for MS experiment}

\item{res_define_at}{numeric m/z threshold for using ppm tolerance for MS1 or MS2 match,
for smaller m/z values, using the tolerance of ppm.ms1 * res.defineat to mathing experimental
and librarial fragments.}

\item{nSlaves}{\code{numeric} Number of threads to be used for multiprocessing}

\item{BPPARAM}{\code{BiocParallelParam or NULL} BPPARAM for multiprocessing data processing}
}
\description{
Params for MS experiment
}
