% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{ExtractIMDataParam}
\alias{ExtractIMDataParam}
\title{parameters for extracting target ion EIC/EIMs}
\usage{
ExtractIMDataParam(
  order_column = "target_intensity",
  mz_tol = 20,
  frame_range = 30,
  frame_integration_range = 5,
  mobility_range = 0.1,
  mobility_intgration_range = 0.015,
  min_points = 4,
  min_intensity = 0,
  n_skip = 0,
  interpolate_method = NULL,
  smooth_method = "gaussian",
  snthreshold = 3,
  peak_span_eim = 21,
  peak_span_eic = 7,
  smooth_window_eim = 16,
  smooth_window_eic = 8,
  skip_invalid_eic_peaks = FALSE,
  skip_invalid_eim_peaks = TRUE,
  keep_profile = FALSE,
  rerun = FALSE
)
}
\description{
parameters for extracting target ion EIC/EIMs
}
