% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{FillPeakParam}
\alias{FillPeakParam}
\title{params for filling peaks}
\usage{
FillPeakParam(
  mz_tol = 20,
  frame_integration_range = 5,
  mobility_tol = 0.015,
  mobility_intgration_range = 0.015,
  rerun = FALSE
)
}
\description{
params for filling peaks
}
