% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{GaussianSmoothParam}
\alias{GaussianSmoothParam}
\title{params for gaussian smoothing}
\usage{
GaussianSmoothParam(
  window = getOption("smoother.window"),
  alpha = getOption("smoother.gaussianwindow.alpha")
)
}
\description{
params for gaussian smoothing
}
