% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{GroupDensityParam}
\alias{GroupDensityParam}
\title{parameters for density peak grouping}
\usage{
GroupDensityParam(
  bw = 5,
  mz_bin_size = 0.015,
  mobility_bin_size = 0.015,
  max_features = 50,
  plot_density = FALSE,
  rerun = FALSE
)
}
\description{
parameters for density peak grouping
}
