% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{MatchBetweenRunParam}
\alias{MatchBetweenRunParam}
\title{parameters for gap filling}
\usage{
MatchBetweenRunParam(
  mz_tol = 20,
  frame_range = 30,
  frame_integration_range = 5,
  mobility_range = 0.1,
  mobility_intgration_range = 0.015,
  min_points = 4,
  min_intensity = 0,
  n_skip = 0,
  interpolate_method = NULL,
  smooth_method = c("gaussian", "loess"),
  peak_span_eim = 21,
  peak_span_eic = 7,
  smooth_window_eim = 16,
  smooth_window_eic = 8,
  keep_profile = FALSE,
  filter_outlier_peaks = TRUE,
  allowed_mobility_shift = 0.015,
  allowed_rt_shift = 10,
  rerun = FALSE
)
}
\description{
parameters for gap filling
}
