\name{Met4DX-package}
\alias{Met4DX-package}
\alias{Met4DX}
\docType{package}
\title{
\packageTitle{Met4DX}
}
\description{
\packageDescription{Met4DX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Met4DX}
\packageIndices{Met4DX}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{Met4DX}

Maintainer: \packageMaintainer{Met4DX}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
