% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Met4DX}
\alias{Met4DX}
\title{MetAnalyzer2_QE}
\description{
Processing MS data
}
\details{
This package processes MS data for peak detection and MSMS spectra assignment
to generate feature table and corresponding MSMS spectra.
}
\author{
Mingdu Luo(luomd@sioc.ac.cn), Yandong Yin (\email{yinyandong@sioc.ac.cn})
}
