% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{QueryTimsDataParam}
\alias{QueryTimsDataParam}
\title{parameters for querying TIMS data}
\usage{
QueryTimsDataParam(opentims_thread = 2, rerun = FALSE)
}
\description{
parameters for querying TIMS data
}
