% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{ReadSpectraParam}
\alias{ReadSpectraParam}
\title{Parameters for reading spectra}
\usage{
ReadSpectraParam(
  intensity_from = c("pepmass", "ms2_intensity"),
  include_precursor = TRUE,
  mz_tol = 20,
  denoise = TRUE,
  int_threshold_abs = 30,
  int_threshold_rel = 0.01,
  rerun = FALSE
)
}
\arguments{
\item{intensity_from}{\code{character} where the precursor intensities were obtained}

\item{include_precursor}{\code{logical} if include precursor ion in MSMS spectra}

\item{mz_tol}{\code{numeric} m/z tolerance for defining a precursor ion}

\item{denoise}{\code{logical} if denoising MSMS spectra}

\item{int_threshold_abs}{\code{numeric} minimal absolute intensity to be kept in MSMS spectra}

\item{int_threshold_rel}{\code{numeric} minimal relative intensity to be kept in MSMS spectra}

\item{rerun}{\code{logical} if rerun the process}
}
\description{
Parameters for reading spectra
}
