% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{SearchParam}
\alias{SearchParam}
\title{Searching spectra Parameters}
\usage{
SearchParam(
  ppm = 20,
  scoreRT = TRUE,
  toleranceRT = c(30, 60),
  scoreCCS = TRUE,
  toleranceCCS = c(2, 4),
  typeCCS = c("percentage", "absolute"),
  adductIncluded = NULL,
  adductExcluded = NULL,
  adductFile = NULL,
  classIncluded = NULL,
  classExcluded = NULL,
  useMS1ResDefine = TRUE,
  updateRefMZ = TRUE,
  resDefineAt = 200
)
}
\arguments{
\item{ppm}{\code{numeric} ppm tolerance for searching precursor mz}

\item{scoreRT}{\code{logical} if comparing RT}

\item{toleranceRT}{\code{numeric (2)} c(Tmin, Tmax), RT range for trapezoidal score.
\itemize{
   \item Tmin: Topline for trapezoidal score
   \item Tmax: Baseline for trapezoidal score
 }}

\item{scoreCCS}{\code{logical} if comparing CCS}

\item{toleranceCCS}{\code{numeric (2)} c(Cmin, Cmax), CCS range for trapezoidal
 score.
\itemize{
   \item Cmin: Topline for trapezoidal score
   \item Cmax: Baseline for trapezoidal score
 }}

\item{typeCCS}{\code{character} CCS tolerance type, either "percentage" or
"absolute"}

\item{adductIncluded}{\code{character} adduct types to be included in MS1 match}

\item{adductFile}{\code{character} file path for user provided adduct table}

\item{classIncluded}{\code{character} compound classes to be included in MS1 match}

\item{useMS1ResDefine}{\code{logical} if use resDefineAt when calculating precursor m/z tolerance}

\item{updateRefMZ}{\code{logical} if update reference precursor m/z with it's corresponding
adduct type}

\item{resDefineAt}{\code{numeric} m/z (Da) value for resolution definition}

\item{adductExclude}{\code{character} adduct types to be excluded in MS1 match}

\item{classExclude}{\code{character} compound classes to be excluded in MS1 match}
}
\description{
Parameters for searching experimentall related SpectraData from refrence
SpectraData
}
